+++ src/main/java/com/beastlymc/triptimize/dto/request/RegisterRequest.java
@@ -1,6 +1,7 @@
-package com.beastlymc.triptimize.auth;
+package com.beastlymc.triptimize.dto.request;
 
 import java.sql.Date;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
