+++ src/main/java/com/beastlymc/triptimize/model/account/profile/Location.java
@@ -1,4 +1,4 @@
-package com.beastlymc.triptimize.user.profile;
+package com.beastlymc.triptimize.model.account.profile;
 
 import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
